Expense Mangement System/requirements.txt
# requirements

fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.10.4
mysql-connector-python==9.2.0
python-dotenv==1.0.1
streamlit == 1.43.2
requests==2.32.3
pandas==2.2.3
pytest == 7.4.4
pytest-mock == 3.14.0
